Split out migration code
